requirements.txt
pandas-profiling
shap
streamlit
kaggle
xgboost
streamlit-toggle-switch
replicate